tenant-onboarding-portal/pyproject.toml
[project]
name = "tenant-onboarding-portal"
version = "0.1.0"
description = "AI Services Hub – Tenant Onboarding Portal"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115,<1",
    "uvicorn[standard]>=0.34,<1",
    "jinja2>=3.1,<4",
    "python-multipart>=0.0.18,<1",
    "pydantic>=2.10,<3",
    "pydantic-settings>=2.7,<3",
    "azure-data-tables>=12.6,<13",
    "azure-identity>=1.19,<2",
    "authlib>=1.4,<2",
    "httpx>=0.28,<1",
    "itsdangerous>=2.2,<3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8,<9",
    "pytest-asyncio>=0.24,<1",
    "httpx>=0.28,<1",
    "ruff>=0.8,<1",
]

[tool.ruff]
target-version = "py313"
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
